pyproject.toml
[tool.poetry]
name = "yava"
version = "0.0.1"
description = "YAVA - attempt to create generic report generator"
authors = ["Boris Tvaroska <boris@tvaroska.sk>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
vertexai = "1.64.0"
langchain-core = "0.2.38"
langchain-google-vertexai = "1.0.10"
anthropic = {version = "0.34.1", extras = ["vertexai"]}
langgraph = "0.2.14"
datasets = "^2.21.0"
gcsfs = "^2024.6.1"
promptgit = {git = "git@github.com:tvaroska/promptgit.git"}
litellm = "1.44.22"
langchain-community = "0.2.16"

[tool.poetry.group.dev.dependencies]
jupyterlab = ">=4.2.4"
ipywidgets = ">=8.1.3"
black = ">=24.8.0"
isort = ">=5.13.2"
pytest = ">=8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
